HellLord2 added
1. AI changed from Byte to ushort;
2. Chiefsword AI: fireWall damage function fixed
3. mod HellLord2 added with AI

+++ Client/MirObjects/MapObject.cs
@@ -65,7 +65,7 @@ namespace Client.MirObjects
         public bool Dead, Hidden, SitDown, Sneaking;
         public PoisonType Poison;
         public long DeadTime;
-        public byte AI;
+        public ushort AI;
         public bool InTrapRock;
         public int JumpDistance;
